+++ bin/node/rpc/Cargo.toml
@@ -17,6 +17,7 @@ pallet-transaction-payment-rpc = { version = "2.0.0", path = "../../../frame/tra
 substrate-frame-rpc-system = { version = "2.0.0", path = "../../../utils/frame/rpc/system" }
 sp-transaction-pool = { version = "2.0.0", path = "../../../primitives/transaction-pool" }
 sc-consensus-babe = { version = "0.8", path = "../../../client/consensus/babe" }
+sc-consensus-babe-rpc = { version = "0.8", path = "../../../client/consensus/babe/rpc" }
 sp-consensus-babe = { version = "0.8", path = "../../../primitives/consensus/babe" }
 sc-keystore = { version = "2.0.0", path = "../../../client/keystore" }
 sc-consensus-epochs = { version = "0.8", path = "../../../client/consensus/epochs" }
