+++ bin/node/runtime/src/lib.rs
@@ -82,7 +82,7 @@ pub const VERSION: RuntimeVersion = RuntimeVersion {
 	// and set impl_version to 0. If only runtime
 	// implementation changes and behavior does not, then leave spec_version as
 	// is and increment impl_version.
-	spec_version: 217,
+	spec_version: 219,
 	impl_version: 3,
 	apis: RUNTIME_API_VERSIONS,
 };
