+++ client/consensus/babe/src/authorship.rs
@@ -144,7 +144,7 @@ fn claim_secondary_slot(
 /// a primary VRF based slot. If we are not able to claim it, then if we have
 /// secondary slots enabled for the given epoch, we will fallback to trying to
 /// claim a secondary slot.
-pub(super) fn claim_slot(
+pub fn claim_slot(
 	slot_number: SlotNumber,
 	epoch: &Epoch,
 	config: &BabeConfiguration,
