+++ client/consensus/babe/src/lib.rs
@@ -119,10 +119,9 @@ use sp_api::ApiExt;
 
 mod aux_schema;
 mod verification;
-mod authorship;
-pub mod rpc;
+pub mod authorship;
 #[cfg(test)]
-mod tests;
+pub mod tests;
 
 /// BABE epoch information
 #[derive(Decode, Encode, Default, PartialEq, Eq, Clone, Debug)]
