+++ client/consensus/babe/src/tests.rs
@@ -28,8 +28,6 @@ use sp_consensus::{
 	NoNetwork as DummyOracle, Proposal, RecordProof,
 	import_queue::{BoxBlockImport, BoxJustificationImport, BoxFinalityProofImport},
 };
-use sc_keystore::{KeyStorePtr, Store};
-use sp_application_crypto::AppPair;
 use sc_network_test::*;
 use sc_network_test::{Block as TestBlock, PeersClient};
 use sc_network::config::{BoxFinalityProofRequestBuilder, ProtocolConfig};
@@ -38,9 +36,6 @@ use tokio::runtime::current_thread;
 use sc_client_api::{BlockchainEvents, backend::TransactionFor};
 use log::debug;
 use std::{time::Duration, cell::RefCell};
-use sp_keyring::Ed25519Keyring;
-use rpc::{BabeApi, BabeRPCHandler};
-use jsonrpc_core::IoHandler;
 
 type Item = DigestItem<Hash>;
 
@@ -336,16 +331,6 @@ impl TestNetFactory for BabeTestNet {
 	) {
 		closure(&mut self.peers);
 	}
-}
-
-/// creates keystore backed by a temp file
-fn create_temp_keystore<P: AppPair>(authority: Ed25519Keyring) -> (KeyStorePtr, tempfile::TempDir) {
-	let keystore_path = tempfile::tempdir().expect("Creates keystore path");
-	let keystore = Store::open(keystore_path.path(), None).expect("Creates keystore");
-	keystore.write().insert_ephemeral_from_seed::<P>(&authority.to_seed())
-		.expect("Creates authority key");
-
-	(keystore, keystore_path)
 }
 
 #[test]
@@ -826,26 +811,4 @@ fn verify_slots_are_strictly_increasing() {
 		&mut proposer_factory,
 		&mut block_import,
 	);
-}
-
-#[test]
-fn rpc() {
-	let mut net = BabeTestNet::new(1);
-
-	let peer = net.peer(0);
-	let data = peer.data.as_ref().expect("babe link set up during initialization");
-
-	let client = peer.client().as_full().expect("Only full clients are used in tests").clone();
-	let epoch_changes = data.link.epoch_changes.clone();
-	let config = Config::get_or_compute(&*client).expect("lol");
-	let select_chain = peer.select_chain().expect("Full client has select_chain");
-	let keystore = create_temp_keystore::<AuthorityPair>(Ed25519Keyring::Alice).0;
-	let handler = BabeRPCHandler::new(client.clone(), epoch_changes, keystore, config, select_chain);
-	let mut io = IoHandler::new();
-
-	io.extend_with(BabeApi::to_delegate(handler));
-	let request = r#"{"jsonrpc":"2.0","method":"babe_epochAuthorship","params": [],"id":1}"#;
-	let response = r#"{"jsonrpc":"2.0","result":{"5GrwvaEF5zXb26Fz9rcQpDWS57CtERHpNehXCPcNoHGKutQY":{"primary":[0],"secondary":[1,2,4]}},"id":1}"#;
-
-	assert_eq!(Some(response.into()), io.handle_request_sync(request));
 }
